Move declaration of IVideoCapture class into precomp.hpp, make some methods of IVideoCapture pure virtual
