+++ modules/highgui/src/precomp.hpp
@@ -196,6 +196,20 @@ double cvGetRatioWindow_GTK(const char* name);
 double cvGetOpenGlProp_W32(const char* name);
 double cvGetOpenGlProp_GTK(const char* name);
 
+namespace cv
+{
+    class IVideoCapture
+    {
+    public:
+        virtual ~IVideoCapture() {}
+        virtual double getProperty(int) { return 0; }
+        virtual bool setProperty(int, double) { return 0; }
+        virtual bool grabFrame() = 0;
+        virtual bool retrieveFrame(int, cv::OutputArray) = 0;
+        virtual int getCaptureDomain() { return CAP_ANY; } // Return the type of the capture object: CAP_VFW, etc...
+    };
+};
+
 //for QT
 #if defined (HAVE_QT)
 double cvGetModeWindow_QT(const char* name);
